backend-reporter/api-access-python-server/requirements.txt
cryptography==36.0.1
PyJWT==2.3.0